+++ backend/canisters/notifications_index/api/src/lib.rs
@@ -6,24 +6,31 @@ pub use lifecycle::*;
 pub use queries::*;
 pub use updates::*;
 
-use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{SubscriptionInfo, UserId};
 
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub enum NotificationsIndexEvent {
     SubscriptionAdded(SubscriptionAdded),
     SubscriptionRemoved(SubscriptionRemoved),
     AllSubscriptionsRemoved(UserId),
+    UserBlocked(UserId, UserId),
+    UserUnblocked(UserId, UserId),
 }
 
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub enum UserIndexEvent {
+    UserBlocked(UserId, UserId),
+    UserUnblocked(UserId, UserId),
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct SubscriptionAdded {
     pub user_id: UserId,
     pub subscription: SubscriptionInfo,
 }
 
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct SubscriptionRemoved {
     pub user_id: UserId,
     pub p256dh_key: String,