+++ backend/canisters/notifications_index/api/src/updates/mod.rs
@@ -1,4 +1,5 @@
 pub mod add_notifications_canister;
+pub mod c2c_sync_user_index_events;
 pub mod push_subscription;
 pub mod remove_subscription;
 pub mod remove_subscriptions;