+++ backend/canisters/notifications/CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
-- Add `sender` to notifications to prevent notifications from blocked users ([#7330](https://github.com/open-chat-labs/open-chat/pull/7330))
+- Add `sender` to notifications to support blocking notifications from blocked users ([#7330](https://github.com/open-chat-labs/open-chat/pull/7330))
+- Block notifications from blocked users ([#7333](https://github.com/open-chat-labs/open-chat/pull/7333))
 
 ## [[2.0.1519](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1519-notifications)] - 2024-12-13