+++ backend/canisters/user_index/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Restructure to handle autonomous bots ([#7318](https://github.com/open-chat-labs/open-chat/pull/7318))
+- Sync blocked users to NotificationsIndex ([#7333](https://github.com/open-chat-labs/open-chat/pull/7333))
 
 ## [[2.0.1590](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1590-user_index)] - 2025-01-24